# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anke Sports API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/maintenance/broadcasts": {
      "get": {
        "summary": "List Records",
        "operationId": "list_records_api_v1_maintenance_broadcasts_get",
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Record",
        "operationId": "create_record_api_v1_maintenance_broadcasts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastDraft"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/maintenance/broadcasts/{ident}": {
      "get": {
        "summary": "Get Record",
        "operationId": "get_record_api_v1_maintenance_broadcasts__ident__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Edit Record",
        "operationId": "edit_record_api_v1_maintenance_broadcasts__ident__put",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/maintenance/broadcasts/{ident}/publish": {
      "post": {
        "summary": "Publish",
        "operationId": "publish_api_v1_maintenance_broadcasts__ident__publish_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/maintenance/broadcasts/{ident}/suspend": {
      "post": {
        "summary": "Suspend",
        "operationId": "suspend_api_v1_maintenance_broadcasts__ident__suspend_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/maintenance/broadcasts/{ident}/check": {
      "post": {
        "summary": "Check",
        "operationId": "check_api_v1_maintenance_broadcasts__ident__check_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/maintenance/broadcasts/{ident}/device-evidence": {
      "post": {
        "summary": "Device Evidence",
        "operationId": "device_evidence_api_v1_maintenance_broadcasts__ident__device_evidence_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceEvidence"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public-feed": {
      "get": {
        "summary": "Public Feed Info",
        "operationId": "public_feed_info_api_v1_public_feed_get",
        "parameters": [
          {
            "name": "source_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160,
              "title": "Source Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicFeedView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sources": {
      "get": {
        "summary": "Sources",
        "operationId": "sources_api_v1_sources_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "real",
              "title": "Dataset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/{event_id}": {
      "get": {
        "summary": "Event Detail",
        "operationId": "event_detail_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BroadcastAction": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "reason"
        ],
        "title": "BroadcastAction"
      },
      "BroadcastDecision": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Revision"
          },
          "source_and_event_confirmed": {
            "type": "boolean",
            "title": "Source And Event Confirmed"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Until"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "source_and_event_confirmed",
          "valid_until"
        ],
        "title": "BroadcastDecision"
      },
...
      "BroadcastList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BroadcastView"
            },
            "type": "array",
            "title": "Items"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "has_more"
        ],
        "title": "BroadcastList"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}